+++ server.js
@@ -378,6 +378,6 @@ app.post('/test', function (req, res) {
 
 reset();
 
-app.listen(3000, function () {
-  console.log('Example app listening on port 3000!')
-})
+app.listen(3010, function () {
+    console.log('Example app listening on port 3010!')
+});
